[DDW-1098] - Inspect wallet rewards display and remove actively delegated stake info - fixes
