+++ source/renderer/app/components/staking/delegation-center/DelegationCenterHeader.js
@@ -2,14 +2,12 @@
 import React, { Component, Fragment } from 'react';
 import { observer } from 'mobx-react';
 import { defineMessages, intlShape } from 'react-intl';
-import SVGInline from 'react-svg-inline';
 import { get } from 'lodash';
 import styles from './DelegationCenterHeader.scss';
 import CountdownWidget from '../../widgets/CountdownWidget';
 import type { NextEpoch, TipInfo } from '../../../api/network/types';
-import delimeterIcon from '../../../assets/images/delimeter.inline.svg';
-import delimeterSlashIcon from '../../../assets/images/delimeter-slash.inline.svg';
 import { SLOTS_TOTAL } from '../../../config/epochsConfig';
+import { generateFieldPanel } from './helpers';
 
 const messages = defineMessages({
   epoch: {
@@ -29,7 +27,7 @@ const messages = defineMessages({
   },
   headingLeft: {
     id: 'staking.delegationCenter.headingLeft',
-    defaultMessage: '!!!Cardano epoch {nextEpochStart} starts in',
+    defaultMessage: '!!!Cardano epoch {nextEpochNumber} starts in',
     description: 'Headline for the Delegation center.',
   },
   headingRight: {
@@ -79,65 +77,9 @@ export default class DelegationCenterHeader extends Component<Props> {
       index: number
     ) => (
       <Fragment key={keys[index]}>
-        {this.generateFieldPanel(labels, values, index)}
+        {generateFieldPanel(labels, values, index)}
       </Fragment>
     ));
-  };
-
-  generateFieldPanel = (labels: any, values: any, index: number) => {
-    const value = values[index];
-    const includeSlashDelimeter = index === values.length - 2;
-    const includeDotsDelimeter =
-      !includeSlashDelimeter && index !== values.length - 1;
-    const labelStr = labels[index];
-    const valueStr = value.toString();
-    let zeroValues = '';
-    if (index === 1 && valueStr.length < values[index + 1].toString().length) {
-      const zerosToAdd =
-        parseInt(values[index + 1].toString().length, 10) -
-        parseInt(valueStr.length, 10);
-      switch (zerosToAdd) {
-        case 1:
-          zeroValues = '0';
-          break;
-        case 2:
-          zeroValues = '00';
-          break;
-        case 3:
-          zeroValues = '000';
-          break;
-        case 4:
-          zeroValues = '0000';
-          break;
-        default:
-          break;
-      }
-    }
-
-    return (
-      <div className={styles.fieldPanel}>
-        <div className={styles.left}>
-          <div className={styles.fieldLabel}>{labelStr}</div>
-          <div className={styles.fieldValue}>
-            {zeroValues && <span>{zeroValues}</span>}
-            {valueStr}
-          </div>
-        </div>
-        {includeDotsDelimeter && (
-          <div className={styles.right}>
-            <SVGInline svg={delimeterIcon} className={styles.delimeterIcon} />
-          </div>
-        )}
-        {includeSlashDelimeter && (
-          <div className={styles.right}>
-            <SVGInline
-              svg={delimeterSlashIcon}
-              className={styles.delimeterSlashIcon}
-            />
-          </div>
-        )}
-      </div>
-    );
   };
 
   render() {
@@ -145,11 +87,12 @@ export default class DelegationCenterHeader extends Component<Props> {
     const { networkTip, nextEpoch } = this.props;
     const epoch = get(networkTip, 'epoch', '-');
     const nextEpochStart = get(nextEpoch, 'epochStart', '-');
+    const nextEpochNumber = get(nextEpoch, 'epochNumber', '-');
     const slot = get(networkTip, 'slot', '-');
     const totalSlots = SLOTS_TOTAL;
     const headingFirst = intl.formatMessage(messages.headingRight);
     const headingSecond = intl.formatMessage(messages.headingLeft, {
-      nextEpochStart,
+      nextEpochNumber,
     });
     const description = intl.formatMessage(messages.description);
     const fieldPanels = this.generateCurrentEpochPanels(
@@ -168,7 +111,7 @@ export default class DelegationCenterHeader extends Component<Props> {
                 <div className={styles.epochs}>{fieldPanels}</div>
               </div>
             </div>
-            {nextEpochStart > 0 && (
+            {nextEpochStart.length > 0 && (
               <div className={styles.countdownContainer}>
                 <div className={styles.heading}>{headingSecond}</div>
                 <CountdownWidget nextEpochStart={nextEpochStart} />
